♻️ refactor: Create listing_methods module to assist App class
